agents/lead_generation/requirements.txt
requests>=2.31.0
beautifulsoup4>=4.12.0
pandas>=2.1.0
openpyxl>=3.1.0
python-dotenv>=1.0.0
psycopg[binary]>=3.2.0
